% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_data_set_wrapper.R
\name{add_ms2}
\alias{add_ms2}
\title{Add MS2 Spectras}
\usage{
add_ms2(
  mass_data_set,
  path,
  polarity,
  column,
  ms1.ms2.match.mz.tol = 15,
  ms1.ms2.match.rt.tol = 30
)
}
\arguments{
\item{mass_data_set}{your massdataset object.}

\item{path}{the path to your ms2 files.}

\item{polarity}{the polarity of your mgf file, it can be "negative" or "positive".}

\item{column}{the column type of your data, can choose between "rp" or "hilic".}

\item{ms1.ms2.match.mz.tol}{your mz (mass to charge ratio) tolerance for matching ms1 and ms2.}

\item{ms1.ms2.match.rt.tol}{your retention time tolerance for ms1 and s2 matching.}
}
\description{
A wrapper for the \code{mutate_ms2()} function from mass_data_set. It will match your ms1 features inside of your massdataset object and add ms2 features to it.
}
