% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_database.R
\name{add_references}
\alias{add_references}
\title{Combine Databases}
\usage{
add_references(reference, database_path, file_type)
}
\arguments{
\item{reference}{reference database object.}

\item{database_path}{the index of the data.}

\item{file_type}{the type of file you are using, you can choose between
msp, and hmdb}
}
\value{
prints customized message to the console
}
\description{
Add another database to your reference database.
}
\examples{
reference <- read_msp(mums2_example("PSU-MSMLS.msp"))
add_references(reference, mums2_example("PSU-MSMLS.msp"), "msp")

}
