% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_metrics.R
\name{averaged_subsampled_dissimilarity_object}
\alias{averaged_subsampled_dissimilarity_object}
\title{avgDist with Community object}
\usage{
averaged_subsampled_dissimilarity_object(
  community_object,
  size,
  threshold,
  diversity_index = "bray",
  iterations = 1000
)
}
\description{
avgdist
}
