% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{change_rt_to_seconds_or_minute}
\alias{change_rt_to_seconds_or_minute}
\title{Change RT time to minutes or seconds}
\usage{
change_rt_to_seconds_or_minute(mpactr_object, rt_type = "seconds")
}
\arguments{
\item{mpactr_object}{The object created from \code{import_all_data()}
file.}

\item{rt_type}{how you want to convert your retention time,
your options are minutes, or seconds. defaults to seconds.}
}
\value{
a modified \code{mpactr} object.
}
\description{
This function will change your ms1 peak table rt time to
rt time in seconds or minutes. This modification happens in place
(or by reference), so the \code{mpactr_object} will be updated.
}
\examples{
data <-
   import_all_data(peak_table =
                   mums2::mums2_example("botryllus_pt_small.csv"),
                   meta_data =
                   mums2::mums2_example("meta_data_boryillus.csv"),
                   format = "None")
change_rt_to_seconds_or_minute(data, "minutes")
}
