% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragmentation_tree.R
\name{compute_molecular_formulas2}
\alias{compute_molecular_formulas2}
\title{Compute Molecular formula2}
\usage{
compute_molecular_formulas2(
  matched_data,
  parent_ppm = 3,
  number_of_threads = detectCores() - 1
)
}
\arguments{
\item{matched_data}{your mass_data object generated from \code{ms2_ms1_compare()}}

\item{parent_ppm}{the ppm you wish to generate the candidate
molecular formulas.}

\item{number_of_threads}{the amount of threads we the algorithm will use.}
}
\description{
de novo algorithm for computing molecular formulas. Using fragmentation trees
we are able to generate a resultant molecular formula. To ensure efficient
we are using a greedy heurstic to generate the resultant formula. Although
this may not always result in the correct prediction, it allows us to
efficiently calculate a multitudeof chemical formulas.
}
