% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_metaboscape2mass_dataset.R
\name{convert_metaboscape2mass_dataset}
\alias{convert_metaboscape2mass_dataset}
\title{Create \code{mass_dataset} from Metaboscape feature table}
\usage{
convert_metaboscape2mass_dataset(
  metaboscape_ft,
  sample_info,
  sample_info_note = NULL
)
}
\arguments{
\item{metaboscape_ft}{a Metaboscape feature table as a \code{data.frame} or file
path.}

\item{sample_info}{a \code{data.frame} with, at minimum, the columns \code{sample_id},
\code{class}, \code{group}, and \code{injection.order}.}

\item{sample_info_note}{optional, a \code{data.frame} with two columns \code{name} and
\code{meaning}. \code{name} should contain all column names in \code{sample_info} as
\code{character} \code{meaning} is a \code{character} vector describing the meaning for
the corresponding column.}
}
\value{
A \code{mass_dataset} object.
}
\description{
\code{convert_metaboscape2mass_dataset} takes a \code{data.frame} of sample
metadata (formatted with TidyMass requirements) and a \code{.csv} feature
table exported from Metaboscape to create the S4 \code{mass_dataset}
object (Shen 2023).
}
\details{
This function allows users to import a Metaboscape feature table into
TidyMass structure (\code{mass_dataset}).

To create a \code{mass_dataset} object, we need at least three data frames:
\enumerate{
\item sample_info - a \code{data.frame} with, at minimum, the columns \code{sample_id},
\code{class}, \code{group}, and \code{injection.order}.
\item variable_info - a \code{data.frame} with, at minimum, the columns
\code{variable_id}, \code{mz}, and \code{rt}. The data frame holds our metadata for each
MS1 feature in our feature table, where \code{variable_id} is the MS1 feature
id, and \code{mz} and \code{rt} are the precursor mz and retention time, respectively,
for each MS1 feature. Other feature information can be included, and is
encouraged with the mums package function
\code{convert_metaboscape2mass_dataset()} where any non-sample columns in the
feature table .csv are retained in variable_info.
\item expression_data - a \code{data.frame} where each row is an MS1 feature and
each column is an injection sample.
}

Information for \code{sample_info} of the \code{mass_dataset} object is provided in
the \code{sample_info} argument. To format metadata for input as \code{sample_info}
see https://massdataset.tidymass.org/articles/data_import_and_export.

A Metaboscape feature table is expected for the argument \code{metaboscape_ft},
which  holds information for the \code{variable_info} and \code{expression_data}
slots of the \code{mass_dataset} object. The function extracts all columns
from the feature table that are not mean or absolute intensity values
for groups/samples and stores them as \code{variable_info}. The
\code{expression_data} is then extracted using sample_id from the \code{sample_info}
\code{data.frame}. As a result, the sample_id in \code{sample_info} is expected
to match sample column names in the Metaboscape feature table.
}
\examples{
\dontrun{
# sample metadata
mjb_metadata
# Metaboscape feature table
mjb_metaboscape

mjb_md <- convert_metaboscape2mass_dataset(mjb_metaboscape,
                                           mjb_metadata)}

}
\references{
Shen X (2023). massdataset: massdataset. R package version 1.0.28,
https://github.com/tidymass/massdataset.
}
