% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_metaboscape2mass_dataset.R
\name{convert_mpactr_object_to_mass_data_set}
\alias{convert_mpactr_object_to_mass_data_set}
\title{Conversion mpactr objec to mass data set}
\usage{
convert_mpactr_object_to_mass_data_set(mpactr_object)
}
\arguments{
\item{mpactr_object}{An mpactr object that was generated when you
imported your peak table and meta data.}
}
\description{
This function converts and mpactr object to a mass data set. Since
we use mpactr to import and filter data, we are also able to transform
the object into a mass data set for further usage
}
