% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_structures.R
\name{convert_to_group_averages}
\alias{convert_to_group_averages}
\title{Convert Samples to Group Averages}
\usage{
convert_to_group_averages(matched_data, mpactr_object)
}
\arguments{
\item{matched_data}{your mass data set
object generated from \code{ms2_ms1_compare()}.}

\item{mpactr_object}{The object created from \code{import_all_data()}.}
}
\value{
a \code{mass_data} object using group averages
}
\description{
To account for users measuring there data in triplicates or other
forms of measurement, we have implemented a function that can
transform your matched data object to use group averages instead of
each sample individually.
}
\examples{
data <-
   import_all_data(peak_table =
                   mums2::mums2_example("botryllus_pt_small.csv"),
                   metadata =
                   mums2::mums2_example("boryillus_metadata.csv"),
                   format = "None")



matched_data <- ms2_ms1_compare(mums2_example("botryllus_v2.gnps.mgf"),
 data, 1, 6)

matched_data_avg <- convert_to_group_averages(matched_data,
                                              data)
}
