% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_structures.R
\name{create_community_matrix}
\alias{create_community_matrix}
\title{create community matrix}
\usage{
create_community_matrix(cluster_object)
}
\arguments{
\item{cluster_object}{the result of the \code{cluster_data()} function.}
}
\description{
Takes the shared dataframe from clustur and converts it into a community matrix
}
