% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_object.R
\name{create_community_matrix_object}
\alias{create_community_matrix_object}
\alias{create_community_matrix_object.mass_data}
\alias{create_community_matrix_object.list}
\title{Create Community Matrix}
\usage{
create_community_matrix_object(data)

\method{create_community_matrix_object}{mass_data}(data)

\method{create_community_matrix_object}{list}(data)
}
\arguments{
\item{data}{the result of the \code{cluster_data()} function, or just a mass_data object created from \code{ms2_ms1_compare()}.}
}
\description{
Takes the shared dataframe from clustur or a massdataset and converts it into a community matrix object
}
