% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_structures.R
\name{create_count_table}
\alias{create_count_table}
\title{Create Count Table}
\usage{
create_count_table(peak_table)
}
\arguments{
\item{peak_table}{your peak table data frame, generated from the mpactr object}
}
\description{
Creats a count table based on your peak table
}
