% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_peaks_data}
\alias{create_peaks_data}
\title{Create peaks_data from MS2 spectra}
\usage{
create_peaks_data(spectrum, precursor_mz)
}
\arguments{
\item{spectrum}{A \code{data.frame} containing an MS2 spectrum with columns
\code{mz} and \code{intensity}.}

\item{precursor_mz}{A \code{numeric} value denoting the precursor mz for the
spectrum.}
}
\value{
A \code{list} of class \code{peaks_data} with two named vectors:
\code{precursor_mz} and \code{spectra}. \code{precursor_mz} is a \code{numeric} vector
containing the precursor mz. \code{spectra} is a \code{data.frame} with two
columns: \code{mz} and \code{intensity}.
}
\description{
Create peaks_data from MS2 spectra
}
\details{
\code{create_peaks_data} will create an object of class \code{peaks_data}
to store an MS2 scan and associated metadata (\emph{e.g.,} precursor mz).
}
\examples{
pmz <- 136.062
mz <- c(52.0220, 53.0349, 53.9952, 53.9984, 54.0015, 54.0025, 54.0046,
          54.0098, 54.0243, 54.0264, 54.8659, 55.0564, 55.9345, 55.9355,
          65.0155, 66.0233, 66.0279, 66.0382, 66.8263, 66.8298, 67.0319,
          67.0492, 67.0515, 67.0538, 67.0585, 68.0272, 71.9321, 72.9389,
          72.9413)
int <- c(8, 276, 15, 6, 7, 753, 21, 15, 9, 25, 29, 11, 6, 22, 44, 33,
           5, 3959, 41, 24, 18, 9, 171, 91, 5, 19, 13, 1481, 13)
peaks <- create_peaks_data(data.frame("mz" = mz,
                                      "intensity" = int),
                           pmz)

# example from a file
peaks <- create_peaks_data(read.csv(example("PSUMSMLS_Adenine.csv")),
                           136.0620)
}
