% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_metrics.R
\name{dist_shared}
\alias{dist_shared}
\title{Distance Shared}
\usage{
dist_shared(
  community_object,
  size,
  threshold,
  diversity_index = "bray",
  subsample = TRUE,
  number_of_threads = 1,
  iterations = 100,
  seed = 123
)
}
\arguments{
\item{community_object}{the object created from
the \code{create_community_object()} function.}

\item{size}{the size you wish to rarefy your diversity matrix to.}

\item{threshold}{the threshold you want your species to
reach before it is included in the rarefaction sum.}

\item{diversity_index}{the diversity index you wish
to calculate diversity. You can choose from:
bray, jaccard, soren, hamming, morista, and thetayc.}

\item{subsample}{if true, we will rarefy the data before we run
the diversity calculations. Default is TRUE.}

\item{number_of_threads}{the number of threads you wish to use for this
calculation.}

\item{iterations}{the amount of times you wish to run your calculation.}

\item{seed}{the RNG (random number generator) seed you would like to use.}
}
\value{
a \code{data.frame} object that shows the
dissimilarity between all samples.
}
\description{
Beta diversity is calculation that allows us to calculate
the differences between two samples. We can conduct this analysis
using the created community objects. The available options for
beta diversity are: Bray Curtis, Jaccard Dissimilarity, Sorenson index,
Hamming Distance, Morista-Horn index, and Yue & Clayton measure
of dissimilarity (or thetayc).
}
\examples{
data <-
   import_all_data(peak_table =
                   mums2::mums2_example("botryllus_pt_small.csv"),
                   metadata =
                   mums2::mums2_example("boryillus_metadata.csv"),
                   format = "None")


matched_data <- ms2_ms1_compare(mums2_example("botryllus_v2.gnps.mgf"),
 data, 1, 6)

dist <- dist_ms2(data = matched_data, cutoff = 0.3, precursor_thresh = 2,
 score_params = modified_cosine_params(0.5), min_peaks = 6,
 number_of_threads = 2)

cluster_results <- cluster_data(distance_df = dist,
 ms2_match_data = matched_data,
 cutoff = 0.3, cluster_method = "opticlust")

community_object <- create_community_matrix_object(cluster_results)

dist_shared(community_object = community_object,
 size = 400, threshold = 100, diversity_index = "bray",
 subsample = TRUE, number_of_threads =  1)
}
