% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_cv_params}
\alias{filter_cv_params}
\title{Filter Cv Parameters}
\usage{
filter_cv_params(cv_threshold = NULL, copy_object = FALSE)
}
\arguments{
\item{cv_threshold}{Coefficient of variation threshold.
A lower cv_threshold will result in more stringent filtering and higher
reproducibility. Recommended values between 0.2 - 0.5.}

\item{copy_object}{A \code{boolean} parameter that allows users to return a copied
object instead of modifying the object.}
}
\value{
a \code{list} object of arguments needed to call the given mpactr
function when supplied to the \code{filter_peak_table()} wrapper function.
}
\description{
Creates a list of filter cv arguments
for the \code{filter_peak_table()} function
}
\examples{
filter_cv_params(0.2)
filter_cv_params(0.2)
}
