% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_group_parameters}
\alias{filter_group_parameters}
\title{Filter Group Parameters}
\usage{
filter_group_parameters(
  group_threshold = 0.01,
  group_to_remove,
  remove_ions = TRUE,
  copy_object = FALSE
)
}
\arguments{
\item{group_threshold}{Relative abundance threshold at which to remove ions.
Default = 0.01.}

\item{group_to_remove}{Biological group name to remove ions from.}

\item{remove_ions}{A \code{boolean} parameter. If \code{TRUE} failing ions will be
removed from the peak table. Default = TRUE.}

\item{copy_object}{A \code{boolean} parameter that allows users to return a copied
object instead of modifying the object.}
}
\value{
a \code{list} object of arguments needed to call the given mpactr function when supplied to the
\code{filter_peak_table()} wrapper function.
}
\description{
Creates a list of filter group arguments
for the \code{filter_peak_table()} function
}
