% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_insource_ions_parameters}
\alias{filter_insource_ions_parameters}
\title{Filter Insource Ions Parameters}
\usage{
filter_insource_ions_parameters(cluster_threshold = 0.95, copy_object = FALSE)
}
\arguments{
\item{cluster_threshold}{Cluster threshold for ion deconvolution.
Default = 0.95.}

\item{copy_object}{A \code{boolean} parameter that allows users to return
a copied object instead of modifying the object.}
}
\value{
a \code{list} object of arguments needed to call the given mpactr function when supplied to the
\code{filter_peak_table()} wrapper function.
}
\description{
Creates a list of filter insource ions arguments
for the \code{filter_peak_table()} function
}
