% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_mispicked_ions_parameters}
\alias{filter_mispicked_ions_parameters}
\title{Filter Mispicked Ions Parameters}
\usage{
filter_mispicked_ions_parameters(
  ringwin = 0.5,
  isowin = 0.01,
  trwin = 0.005,
  max_iso_shift = 3,
  merge_peaks = TRUE,
  merge_method = "sum",
  copy_object = FALSE
)
}
\arguments{
\item{ringwin}{Ringing mass window or detector saturation mass window.
Default = 0.5 atomic mass units (AMU).}

\item{isowin}{Isotopic mass window. Default = 0.01 AMU.}

\item{trwin}{A \code{numeric} denoting the retention time threshold for assessing
if ions should be merged. Default = 0.005.}

\item{max_iso_shift}{A \code{numeric}. Default = 3.}

\item{merge_peaks}{A \code{boolean} parameter to determine if peaks found to
belong to the same ion should be merged in the feature table.}

\item{merge_method}{If merge_peaks is TRUE, a method for how similar peaks
should be merged. Can be one of "sum".}

\item{copy_object}{A \code{boolean} parameter that allows users to return a copied
object instead of modifying the object.}
}
\value{
a \code{list} object of arguments needed to call the given mpactr function when supplied to the
\code{filter_peak_table()} wrapper function.
}
\description{
Creates a list of filter mispicked ions arguments
for the \code{filter_peak_table()} function
}
