% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_peak_table}
\alias{filter_peak_table}
\alias{filter_peak_table.filter_mispicked_ions}
\alias{filter_peak_table.filter_group}
\alias{filter_peak_table.filter_cv}
\alias{filter_peak_table.filter_insource_ions}
\title{Filter Peak Table}
\usage{
filter_peak_table(mpactr_object, params)

\method{filter_peak_table}{filter_mispicked_ions}(mpactr_object, params)

\method{filter_peak_table}{filter_group}(mpactr_object, params)

\method{filter_peak_table}{filter_cv}(mpactr_object, params)

\method{filter_peak_table}{filter_insource_ions}(mpactr_object, params)
}
\arguments{
\item{mpactr_object}{the mpactr_object is an object generated from the
\code{import_all_data()} function.
This is how we begin our pipeline.}

\item{params}{the list of arguments generated from calling one of these
functions: \code{filter_mispicked_ions_params()},
\code{filter_group_params()}, \code{filter_cv_params()},
and \code{filter_insource_ions_params()}.}
}
\value{
a \code{mpactr} object that has been filter based on
the supplied parameters.
}
\description{
This function is a wrapper for all of mpactr's filter functions.
When called with a list of parameters that was generated from one of the
following functions, it will call the subsequent filter:
\code{filter_mispicked_ions_params()}, \code{filter_group_params()},
\code{filter_cv_params()}, and \code{filter_insource_ions_params()}.
You can also find more information on these functions in\code{mpactr}
documentation.
}
\examples{
data.table::setDTthreads(threads = 1)
data <-
   import_all_data(peak_table =
                   mums2::mums2_example("botryllus_pt_small.csv"),
                   metadata =
                   mums2::mums2_example("boryillus_metadata.csv"),
                   format = "None")
filtered_data <- data |>
   filter_peak_table(filter_mispicked_ions_params())
}
