% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_ms2.R
\name{fragmentation_tree}
\alias{fragmentation_tree}
\title{frag}
\usage{
fragmentation_tree(list_of_mz_int, parent_mass)
}
\arguments{
\item{list_of_mz_int}{data}

\item{parent_mass}{data}
}
\description{
Clusters the data together
}
