% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_ms2.R
\name{fragmentation_tree2}
\alias{fragmentation_tree2}
\title{frag}
\usage{
fragmentation_tree2(list_of_mz_int, parent_mass, num_threads = detectCores()/2)
}
\arguments{
\item{list_of_mz_int}{data}

\item{parent_mass}{data}

\item{num_threads}{data}
}
\description{
Clusters the data together
}
