% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_structures.R
\name{generate_a_combined_table}
\alias{generate_a_combined_table}
\title{Create a combined table}
\usage{
generate_a_combined_table(
  matched_data,
  annotations = NULL,
  cluster_data = NULL
)
}
\arguments{
\item{matched_data}{massdata object created from \code{ms2_ms1_compare()}}

\item{annotations}{annotations table created from \code{annotate_ms2()}}

\item{cluster_data}{cluster data created from \code{cluster_data()}}
}
\value{
a \code{data.frame} object.
}
\description{
This function will use the generated matched data,
annotations and cluster data, to create a combined dataframe of
all the generated data. It has the ability to create the
dataframe without annotations or clustering data. However, if
annotations are supplied and a feature has more than one annotation,
the data will be returned in long format.
}
\examples{
data <-
   import_all_data(peak_table =
                   mums2::mums2_example("botryllus_pt_small.csv"),
                   metadata =
                   mums2::mums2_example("boryillus_metadata.csv"),
                   format = "None")



matched_data <- ms2_ms1_compare(mums2_example("botryllus_v2.gnps.mgf"),
 data, 1, 6)

dist <- dist_ms2(data = matched_data, cutoff = 0.3, precursor_thresh = 2,
 score_params = modified_cosine_params(0.5), min_peaks = 0,
 number_of_threads = 2)

cluster_results <- cluster_data(distance_df = dist,
 ms2_match_data = matched_data, cutoff = 0.3, cluster_method = "opticlust")

 massbank <- read_msp(mums2_example("MSMS-Neg-CASMI2016.msp"))
 annotations <- annotate_ms2(mass_data = matched_data,
   reference = massbank, scoring_params = modified_cosine_params(0.5),
   ppm = 1.6e3,
   min_score =  0.1, chemical_min_score = 0, number_of_threads = 2)

generate_a_combined_table(matched_data, annotations, cluster_results)
}
