% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_molecular_formula_predictions}
\alias{get_molecular_formula_predictions}
\title{Get Molecular Formula Predictions}
\usage{
get_molecular_formula_predictions(mass_data)
}
\arguments{
\item{mass_data}{The object generated from \code{ms2_ms1_compare()}.}
}
\value{
a \code{character} vector contain all of your predicted molecular formulas.
}
\description{
Returns all of the molecular formula predictions.
}
\examples{
data <-
   import_all_data(peak_table =
                   mums2::mums2_example("botryllus_pt_small.csv"),
                   meta_data =
                   mums2::mums2_example("meta_data_boryillus.csv"),
                   format = "None")

filtered_data <- data |>
   filter_peak_table(filter_mispicked_ions_params()) |>
   filter_peak_table(filter_cv_params(cv_threshold = 0.2)) |>
   filter_peak_table(filter_group_params(group_threshold = 0.1,
                                             "Blanks")) |>
   filter_peak_table(filter_insource_ions_params())


matched_data <- ms2_ms1_compare(mums2_example("botryllus_v2.gnps.mgf"),
 filtered_data, 0.1, 6)
matched_data <- compute_molecular_formulas(matched_data)

get_molecular_formula_predictions(matched_data)

}
