% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_molecular_formula_preds}
\alias{get_molecular_formula_preds}
\title{Get Molecular Formula Predictions}
\usage{
get_molecular_formula_preds(mass_data)
}
\arguments{
\item{mass_data}{The object generated from \code{ms2_ms1_compare()}.}
}
\value{
a \code{character} vector contain all of your predicted molecular
formulas.
}
\description{
Returns all of the molecular formula predictions.
}
\examples{
data <-
   import_all_data(peak_table =
                   mums2::mums2_example("botryllus_pt_small.csv"),
                   metadata =
                   mums2::mums2_example("boryillus_metadata.csv"),
                   format = "None")


matched_data <- ms2_ms1_compare(mums2_example("botryllus_v2.gnps.mgf"),
 data, 0.02, 1)

matched_data <- compute_molecular_formulas(matched_data)

get_molecular_formula_preds(matched_data)

}
