% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_ms1_data}
\alias{get_ms1_data}
\title{Get MS1 Matches}
\usage{
get_ms1_data(mass_data)
}
\arguments{
\item{mass_data}{The object generated from \code{ms2_ms1_compare()}.}
}
\value{
a \code{data.frame} object containing ms1 data.
}
\description{
A getter that returns the generated ms2 data in your \code{mass_data} object.
}
\examples{
data <-
   import_all_data(peak_table =
                   mums2::mums2_example("botryllus_pt_small.csv"),
                   metadata =
                   mums2::mums2_example("boryillus_metadata.csv"),
                   format = "None")

filtered_data <- data |>
   filter_peak_table(filter_mispicked_ions_params()) |>
   filter_peak_table(filter_cv_params(cv_threshold = 0.2)) |>
   filter_peak_table(filter_group_params(group_threshold = 0.1,
                                             "Blanks")) |>
   filter_peak_table(filter_insource_ions_params())


matched_data <- ms2_ms1_compare(mums2_example("botryllus_v2.gnps.mgf"),
 filtered_data, 10, 6)

get_ms1_data(matched_data)

}
