% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_ms2_matches}
\alias{get_ms2_matches}
\title{Get MS2 Matches}
\usage{
get_ms2_matches(mass_data)
}
\arguments{
\item{mass_data}{The object generated from \code{ms2_ms1_compare()}.}
}
\value{
a \code{data.frame} object containing ms2 data.
}
\description{
A getter that returns the generated ms2 data in your \code{mass_data} object.
}
\examples{
data <-
   import_all_data(peak_table =
                   mums2::mums2_example("botryllus_pt_small.csv"),
                   metadata =
                   mums2::mums2_example("boryillus_metadata.csv"),
                   format = "None")


matched_data <- ms2_ms1_compare(mums2_example("botryllus_v2.gnps.mgf"),
 data, 1, 6)

get_ms2_matches(matched_data)

}
