% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_peaks_data}
\alias{get_peaks_data}
\title{Extract MS2 data from multiple classes}
\usage{
get_peaks_data(data, idx)
}
\arguments{
\item{data}{A data structure containing MS2 data. Currently, this
function supports the classes \code{mass_dataset} and \code{gnps_ref}.}

\item{idx}{A \code{numeric} value denoting the position of the scan in \code{data}.}
}
\value{
A \code{list} of class \code{peaks_data} with two named vectors:
\code{precursor_mz} and \code{spectra}. \code{precursor_mz} is a \code{numeric} vector
containing the precursor mz. \code{spectra} is a \code{data.frame} with two
columns: \code{mz} and \code{intensity}.
}
\description{
Extract MS2 data from multiple classes
}
\details{
\code{get_peaks_data} will extract precursor mz and scan (mz, intensity)
information for MS2 data stored in \code{mass_dataset} and \code{gnps_ref}
data classes. \code{mass_dataset} is an S4 object where ms2 metadata
and peaks data are stored in the \code{ms2_data} slot, while \code{gnps_ref}
is a list of reference scans in which metadata is stored in
\code{gnps_ref[[i]]$info} and peaks data are stored in \code{gnps_ref[[1]]$spec}.

Regardless of \code{data} class type, the position of the desired ms2
scan (\code{idx}) is used to access a single scan.
}
\examples{
# for class "mass_dataset"
\dontrun{
dat <- tidyMassDemo
peaks <- get_peaks_data(dat, 1)}

# for class "gnps_ref"
\dontrun{
 dat <- psu_msmls
 peaks <- get_peaks_data(dat, 1)}
}
