% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_peaks_table_mpactr}
\alias{get_peaks_table_mpactr}
\title{Get Peak Table}
\usage{
get_peaks_table_mpactr(mpactr_object)
}
\arguments{
\item{mpactr_object}{The object generated from \code{import_all_data()}.}
}
\value{
a \code{data.frame} object contain all of the ms1 data.
}
\description{
A wrapper of the \code{mpactr::get_peak_table()} function. Will return
the generated peak table.
}
\examples{
data <-
   import_all_data(peak_table =
                   mums2::mums2_example("botryllus_pt_small.csv"),
                   meta_data =
                   mums2::mums2_example("meta_data_boryillus.csv"),
                   format = "None")

filtered_data <- data |>
   filter_peak_table(filter_mispicked_ions_params()) |>
   filter_peak_table(filter_cv_params(cv_threshold = 0.2)) |>
   filter_peak_table(filter_group_params(group_threshold = 0.1,
                                             "Blanks")) |>
   filter_peak_table(filter_insource_ions_params())

change_rt_to_seconds_or_minute(filtered_data, "minutes")

matched_data <- ms2_ms1_compare(mums2_example("botryllus_v2.gnps.mgf"),
 filtered_data, 10, 6)

get_peaks_table_mpactr(data)

}
