% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_database.R
\name{get_reference_data}
\alias{get_reference_data}
\title{Get Reference Data}
\usage{
get_reference_data(reference, index)
}
\arguments{
\item{reference}{reference database object.}

\item{index}{the index of the data.}
}
\value{
prints customized message to the console
}
\description{
Will return the data inside the reference object based on the index given.
}
\examples{
reference <- read_msp(mums2_example("PSU-MSMLS.msp"))
get_reference_data(reference, 1)

}
