% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_samples}
\alias{get_samples}
\title{Get Samples}
\usage{
get_samples(mass_data)
}
\arguments{
\item{mass_data}{The object generated from \code{ms2_ms1_compare()}.}
}
\value{
a \code{character} vector contain all of your samples.
}
\description{
Returns a list of your samples found in the metadata file.
}
\examples{
data <-
   import_all_data(peak_table =
                   mums2::mums2_example("botryllus_pt_small.csv"),
                   metadata =
                   mums2::mums2_example("boryillus_metadata.csv"),
                   format = "None")


matched_data <- ms2_ms1_compare(mums2_example("botryllus_v2.gnps.mgf"),
 data, 1, 6)

get_samples(matched_data)

}
