% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_ms2_methods.R
\name{gnps_params}
\alias{gnps_params}
\title{GNPS-like similarity between two MS/MS spectra}
\usage{
gnps_params(frag_tolerance)
}
\arguments{
\item{frag_tolerance}{The mz fragment tolerance threshold for aligning
fragment peaks from two ms2 spectra. GNPS default = 0.5.}
}
\value{
A parameters list for similarity scoring method "gnps"
}
\description{
\code{gnps_params()} generates a parameter list to perform GNPS-like cosine
similarity score calculation between two MS2 spectra.
}
\details{
\code{gnps_params()} will initiate cosine scoring based on the Python code by
Wang et al. (2016), which is currently used for cosine scoring in GNPS, to
calculate similarity between two MS2 spectra. This scoring method will
compare peaks data, apply a square root normalization to peak intensities,
align peaks both with and without correction for mass shifts, and calculate
similarity.
}
\references{
Mingxun Wang, Jeremy J. Carver, Vanessa V. Phelan, Laura M. Sanchez,
Neha Garg, Yao Peng, Don Duy Nguyen et al. "Sharing and community curation
of mass spectrometry data with Global Natural Products Social Molecular
Networking." Nature biotechnology 34, no. 8 (2016): 828. PMID: 27504778
}
