% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_all_data}
\alias{import_all_data}
\title{Import all data}
\usage{
import_all_data(peak_table, meta_data, format)
}
\arguments{
\item{peak_table}{The file path to your feature table
file.}

\item{meta_data}{The file path to your meta_data file or \code{data.frame}.}

\item{format}{The expected exported type of your peak table, can be
one of "Progenesis", "Metaboscape", "None".}
}
\value{
a \code{mpactr} object.
}
\description{
This function is a wrapper for the mpactr import_data function.
It will import your peak table and meta data and create a mpactr_object.
}
\examples{
squid_data <- import_all_data(peak_table = mums2::example("squid_peak_table.csv"), 
                            meta_data = mums2::example("squid_meta_data.csv"), 
                             format = "None")
}
