% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_database.R
\name{length.reference_database}
\alias{length.reference_database}
\title{Reference database length}
\usage{
\method{length}{reference_database}(x)
}
\arguments{
\item{x}{reference database object.}
}
\value{
returns the length of the regerence database
}
\description{
returns the length of the database
}
\examples{
reference <- read_msp(mums2_example("PSU-MSMLS.msp"))
length(reference)

}
