% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{mums2_example}
\alias{mums2_example}
\title{Get file paths for examples}
\usage{
mums2_example(file = NULL)
}
\arguments{
\item{file}{Name of a file. If \code{NULL}, all examples files will be listed.}
}
\value{
A file path to example data stored in the \code{inst/extdata} directory
of the package.
}
\description{
mums2 contains a number of example files in the \code{inst/extdata} directory.
This function makes them accessible in documentation that shows how file
paths are used in function examples.
}
\examples{
mums2_example()

mums2_example("PSUMSMLS_Adenine.csv")
}
