% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_ms.R
\name{normalize_ms}
\alias{normalize_ms}
\title{Normalize ms1 feature table}
\usage{
normalize_ms(.data, method = NULL)
}
\arguments{
\item{.data}{A \code{mass_dataset} object with the slot \code{expression_data}
activated (see \code{\link[massdataset:activate_mass_dataset]{massdataset::activate_mass_dataset()}}). \code{expression_data}
is a \code{data.frame} with feature_id as rows and sample_id and columns.
Values are expected to represent absolute intensity for each feature in
the sample.}

\item{method}{A \code{character} to define which normalization method to apply.
Can be one of \code{c("rel_abund")}.}
}
\value{
A \code{data.frame} with feature_id as rows and sample_id and columns.
Values represent relative absolute intensity for each feature in the sample.
}
\description{
Normalize ms1 feature table
}
\examples{
\dontrun{
library(dplyr)
dat <- tidyMassDemo

# replace NA with 0 (or apply missing value imputation)
dat@expression_data <- dat@expression_data \%>\% replace(is.na(.), 0)

dat_ms1_norm <- dat \%>\%
   massdataset::activate_mass_dataset("expression_data") \%>\%
   normalize_ms("rel_abund")

# without normalization
dat@expression_data[1:5, 1:5]

# with normalization
dat_ms1_norm@expression_data[1:5, 1:5]}
}
