% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ms_data.R
\name{print.annotation}
\alias{print.annotation}
\title{Print Annotation Object}
\usage{
\method{print}{annotation}(x, ...)
}
\arguments{
\item{x}{the object created from the read_hmdb}

\item{...}{other parameters that are included in the print function.}
}
\description{
function for print the annotation object object
}
