% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_object.R
\name{print.community_object}
\alias{print.community_object}
\title{Print Community Object}
\usage{
\method{print}{community_object}(x, ...)
}
\arguments{
\item{x}{the object created from the \code{create_community_object()} function.}

\item{...}{other parameters that are included in the \code{print()} function.}
}
\value{
returns \code{matrix} representation of the community object and
prints it to the screen.
}
\description{
S3 function for printing the community object
}
\examples{
data <-
   import_all_data(peak_table =
                   mums2::mums2_example("botryllus_pt_small.csv"),
                   metadata =
                   mums2::mums2_example("boryillus_metadata.csv"),
                   format = "None")



matched_data <- ms2_ms1_compare(mums2_example("botryllus_v2.gnps.mgf"),
 data, 1, 6)

community_object <- create_community_matrix_object(matched_data)
print(community_object)
}
