% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_database.R
\name{print.reference_database}
\alias{print.reference_database}
\title{Print reference}
\usage{
\method{print}{reference_database}(x, ...)
}
\arguments{
\item{x}{reference database object.}

\item{...}{any extra print arguments you want to include.}
}
\value{
prints customized message to the console
}
\description{
print reference objects.
}
\examples{
reference <- read_msp(mums2_example("PSU-MSMLS.msp"))
print(reference)

}
