% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ms_data.R
\name{read_hmdb}
\alias{read_hmdb}
\title{Read HMDB database}
\usage{
read_hmdb(hmdb_file, ms2_folder)
}
\arguments{
\item{hmdb_file}{the xml hmdb file.}

\item{ms2_folder}{the folder path of your ms2 spectra files.}
}
\value{
a \code{reference_database} object.
}
\description{
This function allows you to create an hmdb database. However
you are required to supply an xml hmdb file and a folder path
that contains all of the ms2 spectras from the hmdb download
page https://www.hmdb.ca/downloads.
}
\examples{
read_msp(mums2_example("massbank_example_data.msp" ))

}
\references{
Wishart DS, Guo A, Oler E, Wang F, Anjum A, Peters H, Dizon R,
Sayeeda Z, Tian S, Lee BL, Berjanskii M, Mah R, Yamamoto M,
Jovel J, Torres-Calzada C, Hiebert-Giesbrecht M, Lui VW,
Varshavi D, Varshavi D, Allen D, Arndt D, Khetarpal N,
Sivakumaran A, Harford K, Sanford S, Yee K, Cao X, Budinski Z,
Liigand J, Zhang L, Zheng J, Mandal R, Karu N,
Dambrova M, Schiöth HB, Greiner R, Gautam V. HMDB 5.0:
the Human Metabolome Database for 2022.
Nucleic Acids Res. 2022 Jan 7;50(D1):D622-D631.
doi: 10.1093/nar/gkab1062. PMID: 34986597; PMCID: PMC8728138.
}
