% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ms_data.R
\name{read_mzml_mzxml}
\alias{read_mzml_mzxml}
\title{Read mzml and mzXML files}
\usage{
read_mzml_mzxml(file)
}
\arguments{
\item{file}{the object path of your mzml/mzXML file}
}
\description{
Reader function for mzml and mzXML files
}
