% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mums2-package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{get_peak_table}
\alias{get_metadata}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{mpactr}{\code{\link[mpactr:get_metadata]{get_metadata()}}, \code{\link[mpactr:get_peak_table]{get_peak_table()}}}
}}

