% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_ms2_methods.R
\name{spec_entropy_params}
\alias{spec_entropy_params}
\title{Entropy similarity between two MS/MS spectra}
\usage{
spec_entropy_params(
  ms2_tolerance_in_da = 0.02,
  ms2_tolerance_in_ppm = -1,
  clean_spectra = TRUE,
  min_mz = 0,
  max_mz = 1000,
  noise_threshold = 0.01,
  max_peak_num = 100,
  weighted = TRUE
)
}
\arguments{
\item{ms2_tolerance_in_da}{MS2 peak tolerance in Da, set to -1 to disable.
Defaults to \code{0.02}.}

\item{ms2_tolerance_in_ppm}{MS2 peak tolerance in ppm, set to -1 to disable.
Defaults to \code{-1}.}

\item{clean_spectra}{Either \code{TRUE} or \code{FALSE} to clean the spectra prior to
calculating similarity. See \code{msentropy::clean_spectrum} for more information.
Defaults to \code{TRUE}.}

\item{min_mz}{\code{numeric}, minimum mz to keep, set to -1 to disable. Defaults
to \code{0}.}

\item{max_mz}{\code{numeric}, maximum mz to keep, set to -1 to disable. Defaults
to \code{1000}.}

\item{noise_threshold}{Background intensity threshold, all peaks with
intensity < noise_threshold * max_intensity are removed. Set to -1 to
disable. Defaults to \code{0.01}.}

\item{max_peak_num}{\code{numeric}, maximum number of peaks to keep for score
calculation. Set to -1 to disable. Defaults to \code{100}.}

\item{weighted}{\code{logical} whether weighted or unweighted entropy similarity
will be calculated. Defaults to \code{TRUE}.}
}
\value{
A parameters \code{list} for similarity scoring method "spectral_entropy"
}
\description{
Calculate spectral entropy similarity between two MS2 spectra
}
\details{
\code{spec_entropy_params()} will initiate spectral entropy similarity scoring via
the \code{msentropy} package (Li et al. 2021). For more information about
parameters, see \code{\link[msentropy:msentropy_similarity]{msentropy::msentropy_similarity()}}.
}
\examples{
spec_entropy_params()
}
\references{
Li, Y., Kind, T., Folz, J. et al. Spectral entropy outperforms MS/MS dot
product similarity for small-molecule compound identification, Nat Methods
18, 1524–1531 (2021). https://doi.org/10.1038/s41592-021-01331-z
}
