% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_metrics.R
\name{alpha_summary}
\alias{alpha_summary}
\title{Alpha Diversity Summary}
\usage{
alpha_summary(
  community_object,
  size,
  threshold,
  diversity_index = c("shannon", "simpson"),
  subsample = TRUE,
  number_of_threads = 1,
  iterations = 100,
  seed = 123
)
}
\arguments{
\item{community_object}{the object created from
the \code{create_community_object()} function.}

\item{size}{the size you wish to rarefy your diversity matrix to.}

\item{threshold}{the threshold you want your
species to reach before it is included in the rarefaction sum.}

\item{diversity_index}{the diversity index you wish to calculate diversity,
the options are shannon, simpson, or richness. You may also compute many
indexes at the same time using a vector (ie. c("shannon", "simpson")).}

\item{subsample}{if true, we will rarefy the data before we
run the diversity calculations. Default is TRUE.}

\item{number_of_threads}{the number of threads you wish to use for this
calculation.}

\item{iterations}{the amount of times you wish to run your calculation.}

\item{seed}{the RNG (random number generator) seed you would like to use.}
}
\value{
a \code{data.frame} object that shows the dissimilarity in samples.
}
\description{
Alpha Diversity calculates the amount of diversity in
a single sample. We can conduct this analysis using your created
community object. We support the use of Shannon and Simpson diversity index.
}
\examples{
data <-
   import_all_data(peak_table =
                   mums2::mums2_example("botryllus_pt_small.csv"),
                   metadata =
                   mums2::mums2_example("boryillus_metadata.csv"),
                   format = "None")



matched_data <- ms2_ms1_compare(mums2_example("botryllus_v2.gnps.mgf"),
 data, 1, 6)

dist <- dist_ms2(data = matched_data, cutoff = 0.3, precursor_thresh = 2,
 score_params = modified_cosine_params(0.5), min_peaks = 0,
 number_of_threads = 2)

cluster_results <- cluster_data(distance_df = dist,
 ms2_match_data = matched_data,
 cutoff = 0.3, cluster_method = "opticlust")

community_object <- create_community_matrix_object(cluster_results)

alpha_summary(community_object = community_object, size = 400,
              threshold = 100,
              diversity_index = c("shannon", "simpson", "richness"),
              subsample = TRUE, iterations = 1, number_of_threads = 1)
}
