% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_metrics.R
\name{alpha_summary}
\alias{alpha_summary}
\title{Alpha Diversity Summary}
\usage{
alpha_summary(
  community_object,
  size,
  threshold,
  diversity_index = "shannon",
  iterations = 1000
)
}
\arguments{
\item{community_object}{the object created from the \code{create_community_object()} function.}

\item{size}{the size you wish to rarefy your diversity matrix to.}

\item{threshold}{the threshold you want your species to reach before it is included
in the rarefaction sum.}

\item{diversity_index}{the diversity index you wish to calculate diversity, the two options are
shannon or simpson.}

\item{iterations}{the amount of times you wish to run your diversity metrics.}
}
\description{
alpha diversity
}
