% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_metrics.R
\name{alpha_summary}
\alias{alpha_summary}
\title{Alpha Diversity Summary}
\usage{
alpha_summary(
  community_object,
  size,
  threshold,
  diversity_index = "shannon",
  subsample = TRUE,
  number_of_threads = detectCores(),
  iterations = 1000,
  seed = 123
)
}
\arguments{
\item{community_object}{the object created from
the \code{create_community_object()} function.}

\item{size}{the size you wish to rarefy your diversity matrix to.}

\item{threshold}{the threshold you want your
species to reach before it is included in the rarefaction sum.}

\item{diversity_index}{the diversity index you wish to calculate diversity,
the two options are shannon or simpson.}

\item{subsample}{if true, we will rarefy the data before we
run the diversity calculations. Default is TRUE.}

\item{number_of_threads}{the amount of threads
you want the calculation to use.}

\item{iterations}{the amount of times you wish to run your calculation.}

\item{seed}{the RNG (random number generator) seed you would like to use.}
}
\value{
a \code{data.frame} object that shows the dissimilarity in samples.
}
\description{
alpha diversity
}
\examples{
data <- import_all_data(peak_table =
                        mums2::mums2_example("full_mix_peak_table_small.csv"),
                        meta_data =
                        mums2::mums2_example("full_mix_meta_data_small.csv"),
                        format = "Metaboscape")

filtered_data <- data |>
   filter_peak_table(filter_mispicked_ions_parameters()) |>
   filter_peak_table(filter_cv_parameters(cv_threshold = 0.2)) |>
   filter_peak_table(filter_group_parameters(group_threshold = 0.1,
                                             "Blanks")) |>
   filter_peak_table(filter_insource_ions_parameters())

 change_rt_to_seconds_or_minutes(filtered_data, "minutes")

matched_data <- ms2_ms1_compare(mums2_example("full_mix_ms2_small.mgf"),
 filtered_data, 2, 6)
dist <- dist_ms2(data = matched_data, cutoff = 0.3, precursor_thresh = 2,
 score_params = gnps_params(0.5), min_peaks = 0)

cluster_results <- cluster_data(distance_df = dist,
 ms2_match_data = matched_data,
 cutoff = 0.3, cluster_method = "opticlust")

community_object <- create_community_matrix_object(cluster_results)

alpha_summary(community_object, 4000, 100, "shannon", TRUE, iterations = 1)
}
