% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_ms2.R
\name{annotate_ms2}
\alias{annotate_ms2}
\title{Annotate LC-MS/MS features}
\usage{
annotate_ms2(mass_data, reference, scoring_params,
                    ppm, min_score,
                    chemical_min_score,
                    cluster_data = NULL, min_peaks = 0,
                    number_of_threads = 1)
}
\arguments{
\item{mass_data}{The object generated from \code{ms2_ms1_compare()}.}

\item{reference}{Your reference database generated from the
\code{read_msp()} function.
We currently only support msp files.}

\item{scoring_params}{Parameters for scoring method to be applied.
This can be either \code{modified_cosine_params()} or \code{spec_entropy_params()}.}

\item{ppm}{Parts per million. MS2 scans with a
difference in ppm less than or equal to this value will be scored.}

\item{min_score}{Similarity score threshold to determine a match for
annotation. Comparisons with scores below this value will not be reported.}

\item{chemical_min_score}{data2}

\item{cluster_data}{the cluster object that was generated by
\code{cluster_data()}. Can also remain NULL if you wish to annotate the data
without omu information.}

\item{min_peaks}{the minimum number of peaks that need to be present before
you compare the ms2 spectra.}

\item{number_of_threads}{the number of threads you wish to use for this
calculation. Defaults to the number of threads on your computer.}
}
\value{
A \code{data.frame} with all comparisons with scores above the threshold.
Information for the query scan include \code{query_ms1_id} (the variable_id
for features in expression_data of the \code{mass_data} object)
\code{"query_ms2_id"} (the \code{ms2_spectrum_id} in the \code{query}
object), \code{"query_mz"} (the precursor mz for the scan), and \code{"query_rt"}
(the retention time for the scan). \code{query_mz} and \code{query_rt} are derived
from the ms2 matches data. A column (\verb{"ref_idx}) is included to report the
location for the matching reference molecule in \code{"reference"}. Scores
are reported in the \code{"score"} column. \code{query_formula} and
\code{chemical_similarity}
are also reported. Annotation information is returned
given the information provided in the reference used as input.

a \code{data.frame} object containing annotations
}
\description{
Annotate query LC-MS/MS features in a \code{mass_data} object given a
reference list.

\code{annotate_ms2()} allows for annotation of mass spectrometry features.
Similarity between query and reference level 2 spectra are determined via
spectral scoring methods. Currently scoring methods \code{"gnps"} and
\code{"spectral_entropy"} are supported. The scoring method is specified by the
\code{score_params} argument. \code{score_params} is a list of parameters for the
chosen scoring method. Parameters for "gnps" and "spectral_entropy" can be
created with functions \code{\link[=modified_cosine_params]{modified_cosine_params()}} and
\code{\link[=spec_entropy_params]{spec_entropy_params()}}, respectively. If you are using an hmdb
reference database, or a database that does not contain a precursormz
please ensure that you expand your ppm to account for the difference.
}
\examples{
data <-
   import_all_data(peak_table =
                   mums2::mums2_example("botryllus_pt_small.csv"),
                   metadata =
                   mums2::mums2_example("boryillus_metadata.csv"),
                   format = "None")


matched_data <- ms2_ms1_compare(mums2_example("botryllus_v2.gnps.mgf"),
 data, 1, 6)
 massbank <- read_msp(mums2_example("MSMS-Neg-Respect.msp"))
 annotations <- annotate_ms2(mass_data = matched_data,
   reference = massbank, scoring_params = modified_cosine_params(0.5),
   ppm = 1.6e3,
   min_score =  0.5, chemical_min_score = 0,
   number_of_threads = 2)

}
