% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_ms2.R
\name{annotate_ms2}
\alias{annotate_ms2}
\alias{annotate_ms2.mass_data}
\title{Annotate LC-MS/MS features}
\usage{
annotate_ms2(mass_data, reference, scoring_params, 
                         precursor_tolerance, min_score, 
                         chemical_min_score, cluster_data = NULL, min_peaks = 0)

\method{annotate_ms2}{mass_data}(
  mass_data,
  reference,
  scoring_params,
  precursor_tolerance,
  min_score,
  chemical_min_score,
  cluster_data = NULL,
  min_peaks = 0
)
}
\arguments{
\item{mass_data}{The object generated from \code{ms2_ms1_compare()}.}

\item{reference}{Your reference database generated from the \code{read_msp()} function.
We currently only support msp files.}

\item{scoring_params}{Parameters for scoring method to be applied. This can be either
\code{gnps_params()} or \code{spec_entropy_params()}.}

\item{precursor_tolerance}{Precursor mz tolerance. MS2 scans with a
difference in precursor mz less than or equal to this value will be scored.}

\item{min_score}{Similarity score threshold to determine a match for
annotation. Comparisons with scores below this value will not be reported.}

\item{chemical_min_score}{data2}

\item{cluster_data}{the cluster object that was generated by \code{cluster_data()}.
Can also remain NULL if you wish to annotate the data without omu information.}

\item{min_peaks}{the minimum number of peaks that need to be present before
you compare the ms2 spectra.}
}
\value{
A \code{data.frame} with all comparisons with scores above the threshold.
Information for the query scan include \code{query_ms1_id} (the variable_id
for features in expression_data of the \code{mass_data} object)
\code{"query_ms2_id"} (the \code{ms2_spectrum_id} in the \code{query}
object), \code{"query_mz"} (the precursor mz for the scan), and \code{"query_rt"}
(the retention time for the scan). \code{query_mz} and \code{query_rt} are derived
from the ms2 matches data. A column (\verb{"ref_idx}) is included to report the
location for the matching reference molecule in \code{"reference"}. Scores
are reported in the \code{"score"} column. \code{query_formula} and \code{chemical_similarity}
are also reported. Annotation information is returned
given the information provided in the reference used as input.
}
\description{
Annotate query LC-MS/MS features in a \code{mass_data} object given a
reference list.

\code{annotate_ms2()} allows for annotation of mass spectrometry features.
Similaritity between query and refrence level 2 spectra are determined via
spectral scoring methods. Currently scoring methods \code{"gnps"} and
\code{"spectral_entropy"} are supported. The scoring method is specified by the
\code{score_params} argument. \code{score_params} is a list of parameters for the
chosen scoring method. Parameters for "gnps" and "spectral_entropy" can be
created with notur functions \code{\link[=gnps_params]{gnps_params()}} and \code{\link[=spec_entropy_params]{spec_entropy_params()}},
respectively.
}
\examples{
squid_data <- import_all_data(peak_table = mums2::mums2_example("squid_peak_table.csv"), 
                            meta_data = mums2::mums2_example("squid_meta_data.csv"), 
                             format = "None")

squid_filter <- squid_data |>
   filter_peak_table(filter_mispicked_ions_parameters()) |>
   filter_peak_table(filter_cv_parameters(cv_threshold = 0.2)) |>
   filter_peak_table(filter_group_parameters(group_threshold = 0.1, "Blanks")) |>
   filter_peak_table(filter_insource_ions_parameters())


 matched_data <- ms2_ms1_compare(mums2_example("12152023_Coculture_with_new_JC1.gnps.mgf"),
 squid_filter, 2, 6)
 psu_msmls <- read_msp(mums2_example("PSU-MSMLS.msp"))
 annotations <- annotate_ms2(mass_data = matched_data, 
   reference = psu_msmls, scoring_params = gnps_params(0.5),
   precursor_tolerance = 1000, 
   min_score =  0.1, chemical_min_score = .1)

}
