% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_metrics.R
\name{averaged_subsampled_dissimilarity}
\alias{averaged_subsampled_dissimilarity}
\title{avgDist}
\usage{
averaged_subsampled_dissimilarity(
  community_matrix,
  size,
  threshold,
  diversity_index = "bray",
  iterations = 1000
)
}
\description{
avgdist
}
