% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster_data}
\alias{cluster_data}
\title{Cluster}
\usage{
cluster_data(distance_df, mass_data_set, cluster_method = "opticlust")
}
\arguments{
\item{distance_df}{a distance df that was generated from the \code{distMs2()} function}

\item{mass_data_set}{your mass data set object}

\item{cluster_method}{a cluster method, there are five methods to choose from:
furthest, nearest, weighted, average, and opticlust. opticlust is the default}
}
\description{
Clusters the data together
}
