% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_database.R
\name{combined_reference_database}
\alias{combined_reference_database}
\title{Combine Databases}
\usage{
combined_reference_database(reference, other_reference)
}
\arguments{
\item{reference}{reference database object.}

\item{other_reference}{your other reference database object}
}
\value{
prints customized message to the console
}
\description{
Add another database to your reference database.
}
\examples{
reference <- read_msp(mums2_example("massbank_example_data.msp"))
reference2 <- read_msp(mums2_example("massbank_example_data.msp"))
combined_reference_database(reference, reference2)

}
