% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_conversion.R
\name{community_object_to_distance_object}
\alias{community_object_to_distance_object}
\title{Convert distance data frame to a distance object}
\usage{
community_object_to_distance_object(distance_data_frame)
}
\arguments{
\item{distance_data_frame}{the object generated from the \code{dist_shared()} function.}
}
\description{
Converts the object generated from \code{dist_shared()} to a
\code{dist} object.
}
