% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragmentation_tree.R
\name{compute_molecular_formulas}
\alias{compute_molecular_formulas}
\title{Compute Molecular formula Other}
\usage{
compute_molecular_formulas(mass_data, parent_ppm = 3, number_of_threads = 1)
}
\arguments{
\item{mass_data}{your mass_data object generated from \code{ms2_ms1_compare()}}

\item{parent_ppm}{the ppm you wish to generate the candidate
molecular formulas.}

\item{number_of_threads}{the number of threads you wish to use for this
calculation.}
}
\value{
your mass_data object with an additional \code{character}
vector of all the predicted formulas.
}
\description{
de novo algorithm for computing molecular formulas. Using fragmentation trees
we are able to generate a resultant molecular formula. To ensure efficient
we are using a greedy heurstic to generate the resultant formula. Although
this may not always result in the correct prediction, it allows us to
efficiently calculate a multitudeof chemical formulas.
}
\examples{
data <-
   import_all_data(peak_table =
                   mums2::mums2_example("botryllus_pt_small.csv"),
                   metadata =
                   mums2::mums2_example("boryillus_metadata.csv"),
                   format = "None")



matched_data <- ms2_ms1_compare(mums2_example("botryllus_v2.gnps.mgf"),
 data, 0.1, 1)
compute_molecular_formulas(matched_data, number_of_threads = 2)
}
\references{
Sebastian Böcker, Florian Rasche, Towards de novo identification of
metabolites by analyzing tandem mass spectra, Bioinformatics, Volume 24,
Issue 16, August 2008, Pages i49–i55,
https://doi.org/10.1093/bioinformatics/btn270
}
