% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragmentation_tree.R
\name{compute_molecular_formulas}
\alias{compute_molecular_formulas}
\title{frag}
\usage{
compute_molecular_formulas(mass_data, num_threads = detectCores())
}
\arguments{
\item{mass_data}{your mass_data object}

\item{num_threads}{data}
}
\description{
Clusters the data together
}
