% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragmentation_tree.R
\name{compute_molecular_formulas}
\alias{compute_molecular_formulas}
\title{Compute Molecular formula}
\usage{
compute_molecular_formulas(
  mass_data,
  parent_ppm = 3,
  num_threads = detectCores()
)
}
\arguments{
\item{mass_data}{your mass_data object generated from \code{ms2_ms1_compare()}}

\item{parent_ppm}{the ppm you wish to generate the candidate molecular formulas.}

\item{num_threads}{the amount of threads we the algorithm will use.}
}
\value{
your mass_data object with an additional \code{character} vector of all the predicted formulas.
}
\description{
De-novo algorithm for computing molecular formulas. Using fragmentation trees we are able to generate
a resultant molecular formula. To ensure efficent we are using a greedy heurstic to generate the resultant formula.
Although this may not always result in the correct prediction, it allows us to efficently calculate a multitude
of chemical formulas.
}
\examples{
squid_data <- import_all_data(peak_table = mums2::mums2_example("squid_peak_table.csv"), 
                            meta_data = mums2::mums2_example("squid_meta_data.csv"), 
                             format = "None")

squid_filter <- squid_data |>
   filter_peak_table(filter_mispicked_ions_parameters()) |>
   filter_peak_table(filter_cv_parameters(cv_threshold = 0.2)) |>
   filter_peak_table(filter_group_parameters(group_threshold = 0.1, "Blanks")) |>
   filter_peak_table(filter_insource_ions_parameters())


matched_data <- ms2_ms1_compare(mums2_example("12152023_Coculture_with_new_JC1.gnps.mgf"),
 squid_filter, 2, 6)
compute_molecular_formulas(matched_data)
}
