% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragmentation_tree.R
\name{compute_molecular_formulas}
\alias{compute_molecular_formulas}
\title{compute_molecular_formulas}
\usage{
compute_molecular_formulas(
  mass_data,
  parent_ppm = 3,
  num_threads = detectCores()
)
}
\arguments{
\item{mass_data}{your mass_data object}

\item{parent_ppm}{ppm}

\item{num_threads}{data}
}
\description{
Clusters the data together
}
