% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_structures.R
\name{convert_to_group_averages}
\alias{convert_to_group_averages}
\title{Convert Samples to Group Averages}
\usage{
convert_to_group_averages(matched_data, mpactr_object)
}
\arguments{
\item{matched_data}{your mass data set
object generated from \code{ms2_ms1_compare()}.}

\item{mpactr_object}{The object created from \code{import_all_data()}.}
}
\value{
a \code{mass_data} object using group averages
}
\description{
To account for users measuring there data in triplicates or other
forms of measurement, we have implemented a function that can
transform your matched data object to use group averages instead of
each sample individually.
}
\examples{
data <-
   import_all_data(peak_table =
                   mums2::mums2_example("full_mix_peak_table_small.csv"),
                   meta_data =
                   mums2::mums2_example("full_mix_meta_data_small.csv"),
                   format = "Metaboscape")

filtered_data <- data |>
   filter_peak_table(filter_mispicked_ions_params()) |>
   filter_peak_table(filter_cv_params(cv_threshold = 0.2)) |>
   filter_peak_table(filter_group_params(group_threshold = 0.1,
                                             "Blanks")) |>
   filter_peak_table(filter_insource_ions_params())

change_rt_to_seconds_or_minute(filtered_data, "minutes")

matched_data <- ms2_ms1_compare(mums2_example("full_mix_ms2_small.mgf"),
 filtered_data, 2, 6)

matched_data_avg <- convert_to_group_averages(matched_data,
                                                      filtered_data)
}
