% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_object.R
\name{create_community_matrix_object}
\alias{create_community_matrix_object}
\title{community}
\usage{
create_community_matrix_object(cluster_object)
}
\description{
Takes the shared dataframe from clustur and converts it into a community matrix object
}
