% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_object.R
\name{create_community_matrix_object}
\alias{create_community_matrix_object}
\alias{create_community_matrix_object.mass_data}
\alias{create_community_matrix_object.list}
\title{Create Community Matrix}
\usage{
create_community_matrix_object(data)

\method{create_community_matrix_object}{mass_data}(data)

\method{create_community_matrix_object}{list}(data)
}
\arguments{
\item{data}{the result of the \code{cluster_data()} function, or just a mass_data object created from \code{ms2_ms1_compare()}.}
}
\value{
a external pointer to an Rcpp object.
}
\description{
Takes you mass_data object and creates a community matrix object. This object is used to compute diversity calculations.
}
\examples{
squid_data <- import_all_data(peak_table = mums2::mums2_example("squid_peak_table.csv"), 
                            meta_data = mums2::mums2_example("squid_meta_data.csv"), 
                             format = "None")

squid_filter <- squid_data |>
   filter_peak_table(filter_mispicked_ions_parameters()) |>
   filter_peak_table(filter_cv_parameters(cv_threshold = 0.2)) |>
   filter_peak_table(filter_group_parameters(group_threshold = 0.1, "Blanks")) |>
   filter_peak_table(filter_insource_ions_parameters())


matched_data <- ms2_ms1_compare(mums2_example("12152023_Coculture_with_new_JC1.gnps.mgf"),
 squid_filter, 2, 6)

dist <- dist_ms2(data = matched_data, cutoff = 0.3, precursor_thresh = 2,
 score_params = gnps_params(0.5), min_peaks = 0)
cluster_results <- cluster_data(distance_df = dist,
 ms2_match_data = matched_data, cutoff = 0.3, cluster_method = "opticlust")
community_object_with_cluster <- create_community_matrix_object(cluster_results)
community_object_mass_data <- create_community_matrix_object(matched_data)

}
