% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_object.R
\name{create_community_matrix_object}
\alias{create_community_matrix_object}
\title{Create Community Matrix}
\usage{
create_community_matrix_object(cluster_object)
}
\arguments{
\item{cluster_object}{the result of the \code{cluster_data()} function.}
}
\description{
Takes the shared dataframe from clustur and converts it into a community matrix object
}
