% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_structures.R
\name{create_count_table}
\alias{create_count_table}
\title{Create Count Table}
\usage{
create_count_table(mass_data_set)
}
\arguments{
\item{mass_data_set}{your mass data set object.}
}
\description{
Creats a count table based with you mass data set.
}
