% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_ms2.R
\name{dist_ms2}
\alias{dist_ms2}
\title{Calculate pairwise distance between MS/MS features.}
\usage{
dist_ms2(data, cutoff, precursor_thresh, score_params)
}
\arguments{
\item{data}{the object generated from \code{ms2_ms1_compare()}.}

\item{cutoff}{The maximum distance value (\code{numeric}) to store a pairwise
comparison. The default of .3 corresponds to a cosine score of .7,
meaning pairs with a score of .7 or higher will be stored in the matrix.}

\item{precursor_thresh}{Precursor mz tolerance. MS2 scans with a
difference in precursor mz less than or equal to this value will be scored.}

\item{score_params}{Parameters for scoring method to be applied.
See \code{\link[=gnps_params]{gnps_params()}} and \code{\link[=spec_entropy_params]{spec_entropy_params()}} for more details.}
}
\value{
A sparse matrix of class \code{"data.frame"}
}
\description{
\code{dist_ms2} calculates and stores all non-zero distance values above
the user defined cutoff (default = 0.3).
}
\details{
This function takes a \code{mass_data} object as input and calculates distance
between ms2 peaks. Currently, MS1 features without MS2 peaks returns
no distance value. Distance can be calculated with method \code{"gnps"}
or \code{"spectral_entropy"}. A sparse matrix is returned.
}
