% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_metrics.R
\name{diversity}
\alias{diversity}
\title{diversity_object}
\usage{
diversity(community_object, diversity_index)
}
\arguments{
\item{community_object}{the object created from the \code{create_community_object()} function.}

\item{diversity_index}{the diversity index you wish to calculate diversity, the three options are
shannon, simpson, and bray.}
}
\description{
diversity
}
