% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_structures.R
\name{generate_a_combined_table}
\alias{generate_a_combined_table}
\title{Create a combined table}
\usage{
generate_a_combined_table(
  matched_data,
  annotations = NULL,
  cluster_data = NULL
)
}
\arguments{
\item{matched_data}{description}

\item{annotations}{annotations}

\item{cluster_data}{cluster}
}
\value{
a \code{data.frame} object.
}
\description{
combined
}
