% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_object.R
\name{get_community_matrix}
\alias{get_community_matrix}
\title{Get Community Matrix}
\usage{
get_community_matrix(community_object)
}
\arguments{
\item{community_object}{the object created from the \code{create_community_object()} function.}
}
\description{
Returns the community \code{matrix} or the data that you used to create the object.
}
