% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_ms2_peaks_data}
\alias{get_ms2_peaks_data}
\title{Get Peaks Data}
\usage{
get_ms2_peaks_data(mass_data)
}
\arguments{
\item{mass_data}{The object generated from \code{ms2_ms1_compare()}.}
}
\value{
a \code{list} object containing peak data.
}
\description{
A getter that will return the peak data of all of the matched specturms.
The peak data is the list of mz/intensities found in the ms2 file.
}
\examples{
data <-
   import_all_data(peak_table =
                   mums2::mums2_example("botryllus_pt_small.csv"),
                   metadata =
                   mums2::mums2_example("boryillus_metadata.csv"),
                   format = "None")

filtered_data <- data |>
   filter_peak_table(filter_mispicked_ions_params()) |>
   filter_peak_table(filter_cv_params(cv_threshold = 0.2)) |>
   filter_peak_table(filter_group_params(group_threshold = 0.1,
                                             "Blanks")) |>
   filter_peak_table(filter_insource_ions_params())


matched_data <- ms2_ms1_compare(mums2_example("botryllus_v2.gnps.mgf"),
 filtered_data, 10, 6)

get_ms2_peaks_data(matched_data)

}
