% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2_ms1_comparison.R
\name{ms2_ms1_compare}
\alias{ms2_ms1_compare}
\title{Match your ms1 spectra to a ms2}
\usage{
ms2_ms1_compare(ms2_data, mpactr_object, mz_tolerance, rt_tolerance)
}
\arguments{
\item{ms2_data}{the created object from \code{read_mzml_mzxml()} or \code{read_mgf()}}

\item{mz_tolerance}{your mass-charge ratio tolerance}

\item{rt_tolerance}{your retention time tolerance}

\item{ms1_peak_table}{your ms1 peak_table that is created from \code{import_all_data()}}
}
\description{
We are matching your ms1 to your supplied ms2 by looking at the difference between the mz and rt
}
