% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2_ms1_comparison.R
\name{ms2_ms1_compare}
\alias{ms2_ms1_compare}
\title{Match your ms1 spectra to a ms2}
\usage{
ms2_ms1_compare(ms2_files, mpactr_object, mz_tolerance, rt_tolerance)
}
\arguments{
\item{ms2_files}{a list of all your mgf, mzml, or mzxml files.}

\item{mpactr_object}{your mpactr object created from \code{import_all_data()}}

\item{mz_tolerance}{your mass-charge ratio tolerance in ppm
(parts per million).}

\item{rt_tolerance}{your retention time tolerance.}
}
\value{
returns a \code{mass_data} object of all of the ms2 and ms1 matches.
}
\description{
We are matching your ms1 to your supplied ms2 by looking at
the difference between the mz and rt.
}
\examples{
data <-
   import_all_data(peak_table =
                   mums2::mums2_example("botryllus_pt_small.csv"),
                   meta_data =
                   mums2::mums2_example("meta_data_boryillus.csv"),
                   format = "None")

filtered_data <- data |>
   filter_peak_table(filter_mispicked_ions_params()) |>
   filter_peak_table(filter_cv_params(cv_threshold = 0.2)) |>
   filter_peak_table(filter_group_params(group_threshold = 0.1,
                                             "Blanks")) |>
   filter_peak_table(filter_insource_ions_params())

change_rt_to_seconds_or_minute(filtered_data, "minutes")

matched_data <- ms2_ms1_compare(mums2_example("botryllus_v2.gnps.mgf"),
 filtered_data, 10, 6)
}
