% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2_ms1_comparison.R
\name{ms2_ms1_compare}
\alias{ms2_ms1_compare}
\title{Match your ms1 spectra to a ms2}
\usage{
ms2_ms1_compare(ms2_files, mpactr_object, mz_tolerance, rt_tolerance)
}
\arguments{
\item{ms2_files}{a list of either \emph{all} mgf files, mzml, or mzxml files.}

\item{mpactr_object}{your mpactr object creatd from \code{import_all_data()}}

\item{mz_tolerance}{your mass-charge ratio tolerance}

\item{rt_tolerance}{your retention time tolerance}
}
\description{
We are matching your ms1 to your supplied ms2 by looking at the difference between the mz and rt.
}
