% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2_ms1_comparison.R
\name{ms2_ms1_compare}
\alias{ms2_ms1_compare}
\title{Match your ms1 spectra to a ms2}
\usage{
ms2_ms1_compare(ms2_files, mpactr_object, mz_tolerance, rt_tolerance)
}
\arguments{
\item{ms2_files}{a list of either \emph{all} mgf files, mzml, or mzxml files.}

\item{mpactr_object}{your mpactr object creatd from \code{import_all_data()}}

\item{mz_tolerance}{your mass-charge ratio tolerance}

\item{rt_tolerance}{your retention time tolerance}
}
\value{
returns a \code{mass_data} object of all of the matches.
}
\description{
We are matching your ms1 to your supplied ms2 by looking at the difference between the mz and rt.
}
\examples{
squid_data <- import_all_data(peak_table = mums2::example("squid_peak_table.csv"), 
                            meta_data = mums2::example("squid_meta_data.csv"), 
                             format = "None")

squid_filter <- squid_data |>
   filter_peak_table(filter_mispicked_ions_parameters()) |>
   filter_peak_table(filter_cv_parameters(cv_threshold = 0.2, cv_param = "mean")) |>
   filter_peak_table(filter_group_parameters(group_threshold = 0.1, "Blanks")) |>
   filter_peak_table(filter_insource_ions_parameters())


ms2_ms1_compare(example("12152023_Coculture_with_new_JC1.gnps.mgf"), squid_filter, 2, 6)
}
