% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mums2-package.R
\docType{package}
\name{mums2-package}
\alias{mums2}
\alias{mums2-package}
\title{mums2: Mirobial Ecology by Tandem Mass Spectrometry}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mums2/mums2}
  \item Report bugs at \url{https://github.com/mums2/mums2/issues}
}

}
\author{
\strong{Maintainer}: Patrick Schloss \email{pschloss@umich.edu} (\href{https://orcid.org/0000-0002-6935-4275}{ORCID})

Authors:
\itemize{
  \item Allison Mason \email{masonar@umich.edu} (\href{https://orcid.org/0000-0003-1339-1592}{ORCID})
  \item Gregory Johnson \email{grejoh@umich.edu}
}

}
\keyword{internal}
