% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_object.R
\name{print.community_object}
\alias{print.community_object}
\title{Print Community Object}
\usage{
\method{print}{community_object}(x, ...)
}
\arguments{
\item{x}{the object created from the \code{create_community_object()} function.}

\item{...}{other parameters that are included in the print function.}
}
\description{
3 function for print the community object
}
