% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_object.R
\name{print.community_object}
\alias{print.community_object}
\title{Print Community Object}
\usage{
\method{print}{community_object}(x, ...)
}
\arguments{
\item{x}{the object created from the \code{create_community_object()} function.}

\item{...}{other parameters that are included in the print function.}
}
\description{
3 function for print the community object
}
\examples{
data <- import_all_data(peak_table =
                        mums2::mums2_example("full_mix_peak_table_small.csv"),
                        meta_data =
                        mums2::mums2_example("full_mix_meta_data_small.csv"),
                        format = "Metaboscape")

filtered_data <- data |>
   filter_peak_table(filter_mispicked_ions_parameters()) |>
   filter_peak_table(filter_cv_parameters(cv_threshold = 0.2)) |>
   filter_peak_table(filter_group_parameters(group_threshold = 0.1,
                                             "Blanks")) |>
   filter_peak_table(filter_insource_ions_parameters())

change_rt_to_seconds_or_minutes(filtered_data, "minutes")

matched_data <- ms2_ms1_compare(mums2_example("full_mix_ms2_small.mgf"),
 filtered_data, 2, 6)

dist <- dist_ms2(data = matched_data, cutoff = 0.3, precursor_thresh = 2,
 score_params = gnps_params(0.5), min_peaks = 0)

cluster_results <- cluster_data(distance_df = dist,
 ms2_match_data = matched_data, cutoff = 0.3, cluster_method = "opticlust")

community_object <- create_community_matrix_object(cluster_results)
print(community_object)

}
