% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefy_ms.R
\name{rarefy_ms}
\alias{rarefy_ms}
\title{Rarefy MS1 Feature Table}
\usage{
rarefy_ms(community_object, size, threshold, number_of_threads = 1, seed = 123)
}
\arguments{
\item{community_object}{A \code{community_object}}

\item{size}{The desired total sample intensity to subsample to.}

\item{threshold}{The individual feature threshold. Each subsampled feature
must be >= this value to be retained.}

\item{number_of_threads}{the number of threads you wish to use for this
calculation.}

\item{seed}{the RNG (random number generator) seed you would like to use.}
}
\value{
A \code{external_pointer} that references a community
matrix of rarefied feature intensities.

returns a \code{matrix} object that contains your rarefied data.
}
\description{
\code{rarefy_ms()} performs a single subsampling of MS1 features in sample.
Feature intensities are subsampled to the supplied \code{size} and accounts
for intensity thresholds due to machine limits and background noise.
Specifically, features whose abundance falls below the \code{threshold}
after rarefying are removed. This allows for accurate representation
of samples at different dilutions regardless of the desired
submsampling \code{size}.
}
\examples{
data <-
   import_all_data(peak_table =
                   mums2::mums2_example("botryllus_pt_small.csv"),
                   metadata =
                   mums2::mums2_example("boryillus_metadata.csv"),
                   format = "None")


matched_data <- ms2_ms1_compare(mums2_example("botryllus_v2.gnps.mgf"),
 data, 1, 6)

community_object <- create_community_matrix_object(matched_data)
rarefy_ms(community_object, 400, 10)

}
