% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefy_ms.R
\name{rarefy_ms}
\alias{rarefy_ms}
\title{Rarefy MS1 Feature Table}
\usage{
rarefy_ms(data, size, threshold)
}
\arguments{
\item{data}{A \code{tibble} or \code{data.frame} where each row represets an MS1
feature in a sample. The columns \code{mz} and \code{abund} are expected.}

\item{size}{The desired total sample intensity to subsample to.}

\item{threshold}{The individual feature threshold. Each subsampled feature
must be >= this value to be retained.}
}
\value{
A \code{tibble} or \code{data.frame} of rarefied feature intensities.
}
\description{
\code{rarefy_ms()} performs a single subsampling of MS1 features in sample.
Feature intensities are subsampled to the supplied \code{size} and accounts
for intesnity thresholds due to machine limits and background noise.
Specifically, features whose abundance falls below the \code{threshold}
after rarefying are removed. This allows for accurate represetation
of samples at diffrent dilutions regardless of the desired
submsampling \code{size}.
}
\examples{
set.seed(71)

sample1 <- tibble::tibble(
   mz = seq(100, 1000, by = 100),
   abund = round(runif(10, 1000, 5e5)))

rarefy_ms(sample1, 25000, 1000)

}
