% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefy_ms.R
\name{rarefy_ms}
\alias{rarefy_ms}
\title{Rarefy MS1 Feature Table}
\usage{
rarefy_ms(community_object, size, threshold, number_of_threads = detectCores())
}
\arguments{
\item{community_object}{A \code{community_object}}

\item{size}{The desired total sample intensity to subsample to.}

\item{threshold}{The individual feature threshold. Each subsampled feature
must be >= this value to be retained.}

\item{number_of_threads}{the amount of times you wish to run your calculation.
Default is to use all threads.}
}
\value{
A \code{external_pointer} that references a community matrix of rarefied feature intensities.
}
\description{
\code{rarefy_ms()} performs a single subsampling of MS1 features in sample.
Feature intensities are subsampled to the supplied \code{size} and accounts
for intesnity thresholds due to machine limits and background noise.
Specifically, features whose abundance falls below the \code{threshold}
after rarefying are removed. This allows for accurate represetation
of samples at diffrent dilutions regardless of the desired
submsampling \code{size}.
}
\examples{
squid_data <- import_all_data(peak_table = mums2::example("squid_peak_table.csv"), 
                            meta_data = mums2::example("squid_meta_data.csv"), 
                             format = "None")

squid_filter <- squid_data |>
   filter_peak_table(filter_mispicked_ions_parameters()) |>
   filter_peak_table(filter_cv_parameters(cv_threshold = 0.2, cv_param = "mean")) |>
   filter_peak_table(filter_group_parameters(group_threshold = 0.1, "Blanks")) |>
   filter_peak_table(filter_insource_ions_parameters())


matched_data <- ms2_ms1_compare(example("12152023_Coculture_with_new_JC1.gnps.mgf"),
 squid_filter, 2, 6)

dist <- dist_ms2(data = matched_data, cutoff = 0.3, precursor_thresh = 2,
 score_params = gnps_params(0.5), min_peaks = 0)

cluster_results <- cluster_data(distance_df = dist,
 ms2_match_data = matched_data, cutoff = 0.3, cluster_method = "opticlust")

community_object <- create_community_matrix_object(cluster_results)
rarefy_ms(community_object, 4000, 100)

}
