% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ms_data.R
\name{read_hmdb}
\alias{read_hmdb}
\title{Read msp files}
\usage{
read_hmdb(hmdb_file, ms2_folder)
}
\arguments{
\item{hmdb_file}{the xml hmdb file}

\item{ms2_folder}{the folder path of your ms2 files}
}
\value{
a \code{reference_database} object.
}
\description{
Read msp files
}
\examples{
read_msp(mums2_example("PSU-MSMLS.msp"))

}
