% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ms_data.R
\name{read_msp}
\alias{read_msp}
\title{Read msp files}
\usage{
read_msp(msp_file)
}
\arguments{
\item{msp_file}{the file path of your msp file}
}
\value{
a \code{list} object that contains all of the data present in your msp file.
}
\description{
Reader function msp files
}
\examples{
read_msp(mums2_example("PSU-MSMLS.msp"))[[1]]
 
}
